refactor : 패키지 분리 및 접근제어자 변경
