+++ src/main/java/domain/chessboard/ChessBoard.java
@@ -1,7 +1,7 @@
 package domain.chessboard;
 
-import domain.coordinate.Position;
-import domain.coordinate.Route;
+import domain.position.Position;
+import domain.position.Route;
 import domain.piece.Color;
 import domain.piece.Pawn;
 import domain.squarestatus.Empty;
