+++ src/main/java/domain/position/Position.java
@@ -1,4 +1,4 @@
-package domain.coordinate;
+package domain.position;
 
 
 import java.util.HashMap;
