+++ src/main/java/domain/squarestatus/Empty.java
@@ -1,7 +1,7 @@
 package domain.squarestatus;
 
-import domain.coordinate.Position;
-import domain.coordinate.Route;
+import domain.position.Position;
+import domain.position.Route;
 import domain.piece.Color;
 import domain.type.EmptyType;
 import domain.type.Type;
