+++ src/test/java/domain/position/PositionFactoryTest.java
@@ -1,4 +1,4 @@
-package domain.coordinate;
+package domain.position;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
