+++ src/test/java/domain/squarestatus/EmptyTest.java
@@ -1,6 +1,6 @@
 package domain.squarestatus;
 
-import domain.coordinate.Position;
+import domain.position.Position;
 import domain.piece.Color;
 import domain.type.EmptyType;
 import domain.type.PieceType;
